# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FichoChat API",
        "description": "API REST versionnée de FichoChat. Chaque endpoint documente des exemples JSON complets de réponses (succès et erreurs).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Profil privé courant",
                "description": "Retourne l’utilisateur authentifié avec son profil (préférences, avatar, bio).",
                "operationId": "usersMe",
                "responses": {
                    "200": {
                        "description": "Profil privé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersMe200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/profile": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour le profil",
                "description": "Met à jour display_name, bio, locale, theme et avatar. Content-Type: multipart/form-data si avatar fourni, sinon application/json.",
                "operationId": "usersUpdateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "display_name": {
                                        "type": "string",
                                        "example": "Alice Doe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Hello FichoChat"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "fr"
                                    },
                                    "theme": {
                                        "type": "string",
                                        "example": "dark",
                                        "enum": [
                                            "light",
                                            "dark",
                                            "system"
                                        ]
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "display_name": {
                                        "type": "string",
                                        "example": "Alice Doe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Hello FichoChat"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "fr"
                                    },
                                    "theme": {
                                        "type": "string",
                                        "example": "dark",
                                        "enum": [
                                            "light",
                                            "dark",
                                            "system"
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "display_name": "Alice Doe",
                                    "bio": "Hello FichoChat",
                                    "locale": "fr",
                                    "theme": "dark"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersUpdateProfile200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/setup-profile": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Finaliser le profil après vérification email",
                "description": "Définit un username unique et optionnellement un avatar. Obligatoire une fois après confirmation email.",
                "operationId": "usersSetupProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "alice"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "Alice"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "alice"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "Alice"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "username": "alice",
                                    "display_name": "Alice"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil finalisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersUpdateProfile200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Changer le mot de passe",
                "description": "Vérifie le mot de passe actuel. Par défaut, révoque toutes les sessions (refresh tokens).",
                "operationId": "usersChangePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password1!"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword1!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NewPassword1!"
                                    },
                                    "revoke_other_sessions": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "current_password": "Password1!",
                                "password": "NewPassword1!",
                                "password_confirmation": "NewPassword1!",
                                "revoke_other_sessions": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersChangePassword200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorUnauthorized401": {
                "description": "Non authentifié",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object",
                "example": {
                    "success": false,
                    "message": "Unauthenticated."
                }
            },
...
            "UsersMe200": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Operation completed successfully.",
                    "data": {
                        "user": {
                            "id": "019fc7e1-c324-7380-a855-84056ddd751c",
                            "username": "alice",
                            "email": "alice@example.com",
                            "status": "active",
                            "email_verified_at": "2026-08-03T14:00:00.000000Z",
                            "profile": {
                                "display_name": "Alice",
                                "bio": "Bonjour FichoChat",
                                "avatar_url": "http://localhost:8000/storage/avatars/019fc7e1/avatar.jpg",
                                "locale": "fr",
                                "theme": "system",
                                "updated_at": "2026-08-03T15:00:00.000000Z"
                            },
                            "created_at": "2026-08-03T13:48:09.000000Z",
                            "updated_at": "2026-08-03T14:00:00.000000Z"
                        }
                    }
                }
            },
...
            "UsersChangePassword200": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Password updated successfully.",
                    "data": null
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Access Token obtenu via /auth/login ou /auth/register. Header: Authorization: Bearer {access_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}